requirements.txt
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.1
drf-yasg==1.21.7
psycopg2-binary==2.9.9
gunicorn==21.2.0
setuptools>=65.5.1